requirements.txt
openai>=1.0.0
requests>=2.25.0
tavily-python>=0.3.0
streamlit>=1.28.0
python-dotenv>=1.0.0
pydantic>=2.0.0
rich>=13.0.0
